--- spec/models/invitation_form_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-describe InvitationForm do
-  #TODO
-end
